・ExcludePreview実装 ・MultipleEditorの一部処理を別ファイルに分離 ・Foldoutのアイコンが表示されない原因はScrollbarだったっぽい
