+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Multiple.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64dc334f6d8400e49bf14e29e8156c25
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
