+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Multiple/MultipleGUI_VersionFile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae8f1289b28031c468828b8268ab962a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
