+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Single.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 621966aef4e833240add34e93c8706aa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
