+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Single/SingleGUI_DynamicPathVariables.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0443f4d985e2c54b95d30634dc84702
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
