+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Single/SingleGUI_Excludes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c01bb9c4a8719cb488fc4fb4378e310a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
