+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Single/SingleGUI_ExportPackage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 467b2f806a0c1a846b44a4c4ca8445e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
