+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/EditorScript/Single/SingleGUI_VersionFile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9df775a836b62ab4a896d9754fcb1e59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
