+++ Assets/Mizore_Nekoyanagi/Util/MizoresPackageExporter/MizoresPackageExporterConsts.cs
@@ -8,6 +8,7 @@
         public const string EDITOR_PREF_FOLDOUT_OBJECT = EDITOR_PREF_PREFIX + "Foldout_Object";
         public const string EDITOR_PREF_FOLDOUT_REFERENCES = EDITOR_PREF_PREFIX + "Foldout_References";
         public const string EDITOR_PREF_FOLDOUT_EXCLUDES = EDITOR_PREF_PREFIX + "Foldout_Excludes";
+        public const string EDITOR_PREF_FOLDOUT_EXCLUDES_PREVIEW = EDITOR_PREF_PREFIX + "Foldout_ExcludesPreview";
         public const string EDITOR_PREF_FOLDOUT_EXCLUDE_OBJECTS = EDITOR_PREF_PREFIX + "Foldout_ExcludeObjects";
         public const string EDITOR_PREF_FOLDOUT_DYNAMICPATH = EDITOR_PREF_PREFIX + "Foldout_DynamicPath";
         public const string EDITOR_PREF_FOLDOUT_DYNAMICPATH_PREVIEW = EDITOR_PREF_PREFIX + "Foldout_DynamicPathPreview";
